+++ music-app/src/components/Layout/LayoutBar.jsx
@@ -1,0 +1,13 @@
+import { Outlet } from "react-router-dom";
+import Bar from "../Bar/Bar";
+
+export default function LayoutBar() {
+    
+
+    return (
+        <main>
+            <Outlet />
+            <Bar />
+        </main>
+    );
+}
