+++ music-app/src/pages/MyTracks/MyTracks.jsx
@@ -1,7 +1,6 @@
 import { useSelector, useDispatch } from 'react-redux'
 import { useEffect } from 'react'
 import Nav from '../../components/menuNav/MenuNav'
-import Bar from '../../components/Bar/Bar'
 import * as S from './styledMyTracks'
 import CenterBlock from '../../components/centerBlock/CenterBlock'
 import SideBarPlaylist from '../../components/sideBar/sideBarPlaylist'
@@ -27,7 +26,6 @@ function MyTracks() {
         <CenterBlock title="Мои треки" tracks={myTracks} />
         <SideBarPlaylist />
       </S.main>
-      <Bar />
     </S.container>
   )
 }
