+++ music-app/src/pages/main/main.jsx
@@ -2,7 +2,7 @@
 import Nav from '../../components/menuNav/MenuNav'
 import CenterBlock from '../../components/centerBlock/CenterBlock'
 import Sidebar from '../../components/sideBar/sideBar'
-import Bar from '../../components/Bar/Bar'
+
 import * as S from './styledMain'
 
 function Main() {
@@ -13,7 +13,6 @@ function Main() {
         <CenterBlock title="Треки" />
         <Sidebar />
       </S.main>
-      <Bar />
     </S.container>
   )
 }
