+++ music-app/src/pages/playlists/playlist.jsx
@@ -2,7 +2,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { useParams } from 'react-router-dom'
 import Nav from '../../components/menuNav/MenuNav'
-import Bar from '../../components/Bar/Bar'
 import * as S from './styledPlaylist'
 import CenterBlock from '../../components/centerBlock/CenterBlock'
 import SideBarPlaylist from '../../components/sideBar/sideBarPlaylist'
@@ -19,7 +18,7 @@ function Playlist() {
 
   const PL = data.find((p) => p.id === Number(params.id))
   const title = plTitles.find((p) => p.id === Number(params.id)).name
-  console.log(title)
+  
   return (
     <S.container>
       <S.main>
@@ -27,7 +26,6 @@ function Playlist() {
         <CenterBlock title={title} tracks={PL.items} />
         <SideBarPlaylist />
       </S.main>
-      <Bar />
     </S.container>
   )
 }
